+++ adapters/codex/runtime/task-master.cjs
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+'use strict';
+process.env.SPEC_FLOW_HOST_AGENT = '1';
+require('../bin/task-master');